lab3
query & cache added

+++ src/main/java/com/app/weather/service/WeatherService.java
@@ -1,9 +1,12 @@
 package com.app.weather.service;
+
+import com.app.weather.component.CacheComponent;
 import com.app.weather.dto.ConditionDTO;
 import com.app.weather.dto.WeatherDTO;
 import com.app.weather.model.Condition;
 import com.app.weather.model.Weather;
 import com.app.weather.repository.WeatherRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 import java.sql.Timestamp;
 import java.util.List;
@@ -13,20 +16,31 @@ public class WeatherService {
 
     private final WeatherRepository weatherRepository;
     private final ConditionService conditionService;
+    private final CacheComponent cache;
+    private String cacheKey;
 
-    public WeatherService(WeatherRepository weatherRepository, ConditionService conditionService) {
+    public WeatherService(WeatherRepository weatherRepository, ConditionService conditionService, CacheComponent cache) {
         this.weatherRepository = weatherRepository;
         this.conditionService = conditionService;
+        this.cache = cache;
     }
 
+    @Transactional
     public Weather createWeatherWithCondition(WeatherDTO weatherDTO) {
         Weather weather = convertToEntity(weatherDTO);
         weather.setDate(new Timestamp(System.currentTimeMillis()));
 
+        cacheKey = weather.getCity();
+
         // Проверяем, существует ли уже погода для этого города
-        Weather existingWeather = weatherRepository.findByCity(weather.getCity());
+        Weather existingWeather = (Weather) cache.get(cacheKey);
         if (existingWeather != null) {
             return existingWeather;
+        }
+        Weather weatherFromDb = weatherRepository.findByCity(weather.getCity());
+        if (weatherFromDb != null) {
+            cache.put(cacheKey, weatherFromDb);
+            return weatherFromDb;
         }
         // Проверяем, существует ли условие, если нет, то создаем его
         Condition condition = conditionService.getConditionByText(weatherDTO.getCondition().getText());
@@ -40,9 +54,12 @@ public class WeatherService {
         condition.addWeather(weather);
 
         // Создаем новую погоду
-        return weatherRepository.save(weather);
+        Weather savedWeather = weatherRepository.save(weather);
+        cache.put(cacheKey, savedWeather);
+        return savedWeather;
     }
 
+    @Transactional
     public Weather updateWeather(Long id, WeatherDTO weatherDTO) {
         Weather existingWeather = weatherRepository.findById(id).orElse(null);
         if (existingWeather == null) {
@@ -50,6 +67,8 @@ public class WeatherService {
         }
         existingWeather.setDate(new Timestamp(System.currentTimeMillis()));
         existingWeather.setTemperature(weatherDTO.getTemperature());
+
+        cacheKey = existingWeather.getCity();
 
         // Проверяем, существует ли уже погода для этого города
         Weather weatherByCity = weatherRepository.findByCity(weatherDTO.getCity());
@@ -64,21 +83,48 @@ public class WeatherService {
         }
         existingWeather.setCondition(condition);
 
-        return weatherRepository.save(existingWeather);
+        Weather savedWeather = weatherRepository.save(existingWeather);
+        cache.put(cacheKey, savedWeather);
+        return savedWeather;
     }
 
+    @Transactional
     public void deleteWeather(Long id) {
+        Weather weather = weatherRepository.findById(id).orElse(null);
+        if (weather == null) {
+            return;
+        }
+
+        cacheKey = weather.getCity();
+
+        cache.remove(cacheKey);
         weatherRepository.deleteById(id);
     }
 
+    @Transactional
     public Weather getWeatherById(Long id) {
-        return weatherRepository.findById(id).orElse(null);
+        Weather weather = weatherRepository.findById(id).orElse(null);
+        if (weather == null) {
+            return null;
+        }
+
+        cacheKey = weather.getCity();
+
+        cache.put(cacheKey, weather);
+        return weather;
     }
 
+    @Transactional
     public List<Weather> getAllWeathers() {
-        return weatherRepository.findAll();
+        List<Weather> weathers = weatherRepository.findAll();
+        weathers.forEach(weather -> {
+            cacheKey = weather.getCity();
+            cache.put(cacheKey, weather);
+        });
+        return weathers;
     }
 
+    @Transactional
     public WeatherDTO convertToDTO(Weather weather) {
         WeatherDTO dto = new WeatherDTO();
         dto.setId(weather.getId());
@@ -103,4 +149,10 @@ public class WeatherService {
         return weather;
     }
 
+    public List<WeatherDTO> findByTemperature(double temperature) {
+        List<Weather> weathers = weatherRepository.findByTemperature(temperature);
+        return weathers.stream()
+                .map(this::convertToDTO)
+                .toList();
+    }
 }
